requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pillow
pytesseract
python-multipart
openai
sentence-transformers
faiss-cpu
discourse-client
beautifulsoup4